+++ Passepartout/App/iOS/CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## 1.15.0 Beta 2528 (2021-01-26)
+## Unreleased
 
 ### Added
 
@@ -15,6 +15,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Drop hosts restriction in free version ("Unlimited hosts").
+
+### Fixed
+
+- Redundant keychain items.
 
 ## 1.14.0 (2021-01-07)