+++ Passepartout/App/iOS/CHANGELOG.md
@@ -5,11 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## 1.15.3 Beta 2663 (2021-06-26)
+## Unreleased
 
 ### Added
 
 - SurfShark provider.
+- Support for `--compress stub-v2`.
 
 ## 1.15.2 (2021-04-17)