+++ Passepartout/App/iOS/CHANGELOG.md
@@ -5,11 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## 1.16.0 Beta 2683 (2021-07-23)
+## Unreleased
 
 ### Added
 
 - Support for `--scramble xormask`. [#38](https://github.com/passepartoutvpn/passepartout-apple/issues/38)
+- Oeck provider.
 
 ## 1.15.4 (2021-07-21)