+++ Passepartout/App/iOS/CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## 1.17.0 Beta 2810 (2021-10-11)
+## 1.17.0 Beta 2812 (2021-10-11)
 
 ### Fixed